+++ src/main/java/de/fraunhofer/iosb/ilt/frostclient/utils/MqttConfig.java
@@ -61,6 +61,9 @@ public class MqttConfig {
     }
 
     public MqttConnectOptions getOptions() {
+        if (options == null) {
+            options = new MqttConnectOptions();
+        }
         return options;
     }